+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased] - 0000-00-00
 ### Added
 - Initial release.
-- Supported color formats: HEX, RGB, HSL
+- Supported color formats: HEX, RGB, HSL, 256
 
 ### Changed
 ### Deprecated